Handle non-200 response errors in a session (instead of crashing)
Sunstrong Management, who acquired the part of Sunpower's
business dealign with on-prem solar monitoring after the
Sunpower bankruptcy, have pushed a breaking change to the
PVS6 firmware that either disables the installer API or
alters its authentication path; the current effect is that the
exporter crashes on startup because the grid profile response
is empty.

This doesn't fix anything else but it stops the exporter from
crashing on startup and makes the issue more apparent from
logs.